+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `[color]` Added method `RGBA.WithAlpha`
 * `[color]` Improved support for hex colors with alpha
 * `[color]` Use web color representation for `Hex.String`
+* `[color]` Added flag to `Hex.ToWeb` to disable shorthand generation
 * `[color]` Fixed shorthand hex generation for `#FFF`
 * `[color]` Fixed `RGBA` to `Hex` conversion